Updated write_solution_to_file() to allow easier processing of the csv.
Added new optional argument to JuDGE.solve() max_no_int. Which forces an
integer re-solve of the master if no integer solution is found for a given
number of iteration.

Sped up formulations of the JuDGE and DetEq models.

Fixed a minor bug in the branch and price. The upper and lower bounds will
now always be reported correctly.