Several changes have been made.
Added continuous investment variables; this is not well tested and should
only be used if you understand the implicit assumptions.

Updated the advanced.jl example to use the new callback interface in the
current version of JuMP.

Added compatibility with additional datatypes for arrays of variables.

Added a NodeID datatype that is associated with each node in a tree.
This is used to uniquely reference a node in a tree, in such a way that,
we are able to automatically create a set of linear trees (one for each
scenario), while maintaining the data associations, in order to compute the
EVPI.